fix debug & del gqbmm private

Created-by: echo_jia
Commit-by: echo_jia
Merged-by: echo_jia
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #000-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：



See merge request: libohao6/ops-transformer!8

+++ mc2/allto_allv_grouped_mat_mul/op_kernel/mc2_templates/common/a2av_common_tiling.h
@@ -69,8 +69,8 @@ struct TaskTilingInfo {
     uint32_t totalLoopCount;     // 总循环次数
     
     // 通信参数（对应sendCounts和recvCounts）
-    int64_t sendCnt[MAX_EXPERT_SIZE];  // 每个expert的发送计数
-    int64_t recvCnt[MAX_EXPERT_SIZE];  // 每个expert的接收计数
+    int32_t sendCnt[MAX_EXPERT_SIZE];  // 每个expert的发送计数
+    int32_t recvCnt[MAX_EXPERT_SIZE];  // 每个expert的接收计数
 };
 
 }
